+++ Tests/Buildenator.IntegrationTests.Source/Builders/GrandchildEntityNoMoqBuilder.cs
@@ -5,7 +5,7 @@ using Buildenator.IntegrationTests.SharedEntities;
 
 namespace Buildenator.IntegrationTests.Source.Builders
 {
-    [MakeBuilder(typeof(GrandchildEntity))]
+    [MakeBuilder(typeof(GrandchildEntity), defaultStaticCreator: true)]
     [MoqConfiguration(MockingInterfacesStrategy.None)]
     [AutoFixtureConfiguration(strategy: FixtureInterfacesStrategy.None)]
     public partial class GrandchildEntityNoMoqBuilder