+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.2]
+## [3.2.0]
 
 ### Added
 ### Changed
 ### Removed
+
+
+## 3.2.0 - 2021-09-14
+
+### Added
+
+- Static default builder entity method generation
+    - an example of a generated method: 
+    ```public static Entity BuildDefault(int _param1 = default(int), string _param2 = default(string)) { return new Entity(_param1, _param2); }```
 
 
 ## 3.1.2 - 2021-09-14